+++ lib/src/interface_has_ui_display_name.dart
@@ -1,0 +1,3 @@
+abstract class IHasUIDisplayName {
+  String getUiDisplayName();
+}