+++ CHANGELOG.md
@@ -74,4 +74,20 @@ and fix issues
 
 # 0.1.19
 
+<<<<<<< HEAD
 - implementation 1 new component SimpleCard and fix issues
+=======
+- fix issue on RestClientGeneric, not present totalRecords on result list of getAll
+
+# 0.1.20
+
+- fix issue on RestClientGeneric, not present totalRecords on result list of getAll
+
+# 0.1.22
+
+- fix bugs on simple select
+
+# 0.1.23
+
+- fix bugs 
+>>>>>>> 07c8585d5811645923c2fdcccb6f9d9f9ee1456e